requirements.txt
behave==1.2.6
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python~=8.1.0
influxdb~=5.3.1
allure-behave==2.13.2
allure-python-commons==2.13.2
wiremock~=2.6.1